create folder common and move reusable somponents into
