+++ app/contacts/page.jsx
@@ -1,11 +1,11 @@
 "use client";
 import ContextBlock from "./ContextBlock";
-import { Navbar } from "../components/Navbar";
-import Footer from "../components/Footer";
+import {Navbar} from "../common/Navbar";
+import Footer from "../common/Footer";
 import useScreenHandling from "../hooks/useScreenHandling";
 
 const page = () => {
-  const { screen } = useScreenHandling();
+  const {screen} = useScreenHandling();
 
   return (
     <div className="flex overflow-hidden max-h-screen flex-col w-full items-center">
